v0.9.1
Addressed some issues found in the scenario:

++++++++++++++++++++++++SPOILER ALERT+++++++++++++++++++++++++++++++++++++++++++++

FIXED: First Aid kit gives all the time Sedatives and Bandages. It should disappear after one or two search actions.
FIXED: Book token in Student Library doesn't disappear.
FIXED: Students belongings tokens in the student lunge has no spawn text and it doesn't disappear

IGNORING: Visiting the temple shouldn't be possible without visiting first the University.
REASON: This scenario is meant to be played multiple times. It is only
intended for the investigators to go to the temple to try and stop the
evil in the past (a different Time), or to go to the university to try and stop it in the
present (a different Place). This is why the scenario was named A Time and
Place.

IGNORING:
1) Restricted Section door token has 2 options: the fist launches a puzzle to be solved before the door opens, but the second one, it just opens the door without doing anything
2) And the token aside this one is useless (for a non-existing secret passage?)
REASON:
1) This is intended behavior. If the investigators question the
librarian in the cafeteria, she moves to the library and leaves the door
unlocked.
2) The real bug here was that the token should allow passage by a
qouta-based observation check. This token is now removed if passage
is gained to the Restricted Section another way.